:bug: updated cnn test code to fix bug regarding indexing predictions array
Bug-1:
Previously the prediction array was a list of tensors, where each tensor had the shape (BATCH_SIZE, NUM_GENES). These are unshuffled and come in sequence of the slides. This list of tensors was later accessed with the expectation that it would have the shape (NUM_NON_ZERO_SPOTS, NUM_GENES), where NUM_NON_ZERO_SPOTS indicates the total number of spots in the predictions array (which would only have the predictions for the non-zero spots since that is what we train the GAT on). But the predictions array needs to be concatenated for it to have this desired arrangement.

Fix:
Added code to concatenate the list to tensors into one tensor of shape (NUM_NON_ZERO_SPOTS, NUM_GENES). The resulting tensor would have all the predictions for the non-zero spots in sequence of the slides.
For example - if the batch size is 256, and the first slide has 400 spots, the predictions for the first 256 spots of the first slide would be the first batch, and the predictions for the remaining 144 spots for the first slide would be the first 144 elements of the second batch.

Bug-2:
Additionally, while indexing the preds array, the index previously used was [start : start + data['nonzero'][i]]. This was incorrect becase data['nonzero'][i] would provide the numpy array of booleans indexing which spots are non-zero. What we need is rather the total count of non-zero spots, which would be obtained by taking the sum of the boolean array data['nonzero'][i].

Fix:
Replaced `data['nonzero'][i]` with `data['nonzero'][i].sum()`.