:bug: fixed config variables for gnn training
Bug:
Some older config variables were not replaced with the newer config file standard. These included the flag for enabling hierarchical graph construction, as well as the variables indicating the number of spatial and feature space clusters during hierarchical graph construction. The batch_size for the graph dataloader was also dependent on a config variable. Since we intend to process the graphs slide-by-slide, this should be hardcoded to 1.

Fix:
Replaced older config variables with the newer ones. Also hardcoded the graph dataloader batchsize to 1.