:bug: replaced an obsolete config variable and also corrected a bug for indexing tissue_positions dataframe
Bug-1:
The config variable for setting the number of workers in the dataloader was removed but the code for accessing it while creating the graph dataloader was not replaced.

Fix:
Hardcoded the number of workers in the graph dataloader to 4. This is not necessarily the best number.

Bug-2:
The previous format of the MERGE dataset expected a tissue_positions csv file without headers. So it accessed the x and y coordinates for the spots using numerical indices and 4 and 5. The data files were later updated to contain a header row where x and y coordinates would be indicated by `pxl_col_in_fullres` and `pxl_row_in_fullres`.

Fix:
Updated the indexing code in the build_hierarchical_graph function to use the new header names for the respective columns.