:zap: & :bug: improved code and removed graph dataloader bug
Previous code redundantly created a concatenated array of ground truth expressions and patch embeddings for all spots in the dataset.
This would later be used in the GraphDataset initialization to create slide-wise batches.
The counts and patch embeddings are already available in the data dictionary so the entire computation to create the concatenated lists is redundant.
The subsequent indexing to retrieve slide-wise values from the concatenated array is also redundant.

Bug:
Previously there was a single graph dataset and a single graph dataloader consisting of all slides.
This would not work with the subsequent training code as there needs to be separate dataloaders for training and testing.

Fix:
Created separate datasets and dataloaders for training and testing.