:bug: removed old code to train gnn with all graphs
Bug-1:
GNN training expected older version of dataset and did not use any torch dataloader.

Fix:
Replaced the code to accommodate separate train and test dataloaders.

Bug-2:
This stems from the earlier bug. The code expected to process all graphs for all samples in a single step.
This would lead to OOMs if the cohort is sufficiently large.

Fix:
Replaced gnn training and testing code to process the graph from a single slide in one step.
Losses are computed for each such step.