:bug: changed number of works in graph dataloader
Bug-1:
The graph dataloader for training set was harcoded to use 4 workers.
Unfortunately, the collation for multi-worker code does not seem to be functional.
I previously fixed it for the validation graph dataloader, somehow missed the training one.

Fix:
Changed the number of workers to 0 for the training graph dataloader.